@@ -4,8 +4,8 @@ description: Flutter package for widget tree building from html that supports hy
 homepage: https://github.com/daohoangson/flutter_widget_from_html
 
 environment:
-  # flutter_layout_grid@0.10.2
-  flutter: ">=1.14.0 <2.0.0"
+  # flutter_svg@0.18.0
+  flutter: ">=1.18.0-6.0.pre <2.0.0"
   sdk: ">=2.3.0 <3.0.0"
 
 dependencies:
@@ -14,7 +14,7 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_layout_grid: ^0.10.2
-  flutter_svg: ">=0.17.4 <0.19.0"
+  flutter_svg: ^0.18.0
   flutter_widget_from_html_core: ^0.5.0-rc.2020071301
   html: ^0.14.0+3
   url_launcher: ^5.4.10
