@@ -168,14 +168,9 @@ Widget _buildAsyncBuilder(BuildContext _, AsyncSnapshot<Widget> snapshot) =>
 
 Widget _buildBody(WidgetFactory wf, HtmlWidget widget, dom.NodeList domNodes) {
   wf.reset(widget);
-  final tsb = TextStyleBuilders()..enqueue(_tsb, widget.textStyle);
-  final built = HtmlBuilder(domNodes: domNodes, parentTsb: tsb, wf: wf).build();
-  return wf.buildBody(built) ?? widget0;
+  final rootTsb = TextStyleBuilder.root(style: widget.textStyle, wf: wf);
+  final builder = HtmlBuilder(domNodes: domNodes, parentTsb: rootTsb, wf: wf);
+  return wf.buildBody(builder.build()) ?? widget0;
 }
 
 dom.NodeList _parseHtml(String html) => parser.parse(html).body.nodes;
-
-TextStyleHtml _tsb(BuildContext _, TextStyleHtml p, TextStyle style) =>
-    style == null
-        ? p
-        : TextStyleHtml.style(style.inherit ? p.style.merge(style) : style);
