@@ -3,28 +3,34 @@ part of '../core_widget_factory.dart';
 const _kCssPadding = 'padding';
 
 WidgetPlaceholder _paddingInlineAfter(
-  TextStyleBuilders tsb,
+  TextStyleBuilder tsb,
   CssLengthBox box,
 ) =>
-    WidgetPlaceholder(builder: (context, children, _) {
-      final direction = Directionality.of(context);
-      final width = box.right(direction)?.getValue(context, tsb);
-      if (width == null || width <= 0) return null;
+    WidgetPlaceholder<CssLengthBox>(
+      child: Builder(builder: (context) {
+        final direction = Directionality.of(context);
+        final width = box.right(direction)?.getValue(context, tsb);
+        if (width == null || width <= 0) return widget0;
 
-      return [SizedBox(width: width)];
-    });
+        return SizedBox(width: width);
+      }),
+      generator: box,
+    );
 
 WidgetPlaceholder _paddingInlineBefore(
-  TextStyleBuilders tsb,
+  TextStyleBuilder tsb,
   CssLengthBox box,
 ) =>
-    WidgetPlaceholder(builder: (context, children, _) {
-      final direction = Directionality.of(context);
-      final width = box.left(direction)?.getValue(context, tsb);
-      if (width == null || width <= 0) return null;
+    WidgetPlaceholder<CssLengthBox>(
+      child: Builder(builder: (context) {
+        final direction = Directionality.of(context);
+        final width = box.left(direction)?.getValue(context, tsb);
+        if (width == null || width <= 0) return widget0;
 
-      return [SizedBox(width: width)];
-    });
+        return SizedBox(width: width);
+      }),
+      generator: box,
+    );
 
 class _StylePadding {
   final WidgetFactory wf;
@@ -35,7 +41,7 @@ class _StylePadding {
         isBlockElement: false,
         onPieces: (meta, pieces) {
           if (meta.isBlockElement) return pieces;
-          final padding = wf.parseCssPadding(meta);
+          final padding = wf.parseCssLengthBox(meta, _kCssPadding);
           if (padding?.hasLeftOrRight != true) return pieces;
 
           return _wrapTextBits(
@@ -48,37 +54,24 @@ class _StylePadding {
         },
         onWidgets: (meta, widgets) {
           if (widgets?.isNotEmpty != true) return null;
-          final padding = wf.parseCssPadding(meta);
+          final padding = wf.parseCssLengthBox(meta, _kCssPadding);
           if (padding == null) return null;
 
-          final input = _PaddingInput()
-            ..meta = meta
-            ..padding = padding;
-
-          return _listOrNull(wf.buildColumn(widgets)?.wrapWith(_build, input));
+          return _listOrNull(wf
+              .buildColumnPlaceholder(widgets)
+              ?.wrapWith((child) => _build(child, padding, meta.tsb())));
         },
         priority: 9999,
       );
 
-  Iterable<Widget> _build(
-    BuildContext context,
-    Iterable<Widget> children,
-    _PaddingInput input,
-  ) {
-    final direction = Directionality.of(context);
-    final padding = input.padding;
-    final tsb = input.meta.tsb;
-    final top = padding.top?.getValue(context, tsb);
-    final right = padding.right(direction)?.getValue(context, tsb);
-    final bottom = padding.bottom?.getValue(context, tsb);
-    final left = padding.left(direction)?.getValue(context, tsb);
-
-    return _listOrNull(wf.buildPadding(wf.buildColumn(children),
-        EdgeInsets.fromLTRB(left ?? 0, top ?? 0, right ?? 0, bottom ?? 0)));
-  }
-}
-
-class _PaddingInput {
-  NodeMetadata meta;
-  CssLengthBox padding;
+  Widget _build(Widget child, CssLengthBox padding, TextStyleBuilder tsb) =>
+      Builder(builder: (context) {
+        final direction = Directionality.of(context);
+        final t = padding.top?.getValue(context, tsb);
+        final r = padding.right(direction)?.getValue(context, tsb);
+        final b = padding.bottom?.getValue(context, tsb);
+        final l = padding.left(direction)?.getValue(context, tsb);
+        final edgeInsets = EdgeInsets.fromLTRB(l ?? 0, t ?? 0, r ?? 0, b ?? 0);
+        return wf.buildPadding(child, edgeInsets);
+      });
 }
