@@ -24,15 +24,13 @@ class _StyleVerticalAlign {
         },
       );
 
-  BuiltPiece _buildWidgetSpan(BuiltPiece piece, String verticalAlign) {
+  BuiltPiece _buildWidgetSpan(BuiltPiece piece, String value) {
     if (piece.hasWidgets) return piece;
 
-    final text = piece.text;
-    final input = _buildInput(verticalAlign, text.tsb);
-    if (input == null || input.alignment == PlaceholderAlignment.baseline) {
-      return piece;
-    }
+    final alignment = _tryParse(value);
+    if (alignment == null) return piece;
 
+    final text = piece.text;
     final replacement = (text.parent?.sub(text.tsb) ?? TextBits(text.tsb))
       ..detach();
     text.replaceWith(replacement);
@@ -41,72 +39,55 @@ class _StyleVerticalAlign {
     final newPiece = BuiltPiece.text(replacement);
     if (built == null) return newPiece;
 
-    if (input.padding != null) built.wrapWith(_build, input);
-    replacement.add(TextWidget(text, built, alignment: input.alignment));
+    if (value == _kCssVerticalAlignSub || value == _kCssVerticalAlignSuper) {
+      built.wrapWith(
+        (child) => _build(
+          child,
+          EdgeInsets.only(
+            bottom: value == _kCssVerticalAlignSub ? .4 : 0,
+            top: value == _kCssVerticalAlignSuper ? .4 : 0,
+          ),
+          text.tsb,
+        ),
+      );
+    }
+    replacement.add(TextWidget(text, built, alignment: alignment));
 
     return newPiece;
   }
 
-  Iterable<Widget> _build(
-    BuildContext context,
-    Iterable<Widget> widgets,
-    _VerticalAlignInput input,
-  ) {
-    final child = wf.buildColumn(widgets);
-    final fontSize = input.tsb.build(context).style.fontSize;
-    final padding = input.padding;
-    assert(padding != null);
+  Widget _build(Widget child, EdgeInsets padding, TextStyleBuilder tsb) =>
+      Builder(builder: (context) {
+        final fontSize = tsb.build(context).style.fontSize;
 
-    return [
-      Stack(children: <Widget>[
-        wf.buildPadding(
-          Opacity(child: child, opacity: 0),
-          EdgeInsets.only(
-            bottom: fontSize * padding.bottom,
-            top: fontSize * padding.top,
+        return Stack(children: <Widget>[
+          wf.buildPadding(
+            Opacity(child: child, opacity: 0),
+            EdgeInsets.only(
+              bottom: fontSize * padding.bottom,
+              top: fontSize * padding.top,
+            ),
           ),
-        ),
-        Positioned(
-          child: child,
-          bottom: padding.top > 0 ? null : 0,
-          top: padding.bottom > 0 ? null : 0,
-        )
-      ]),
-    ];
+          Positioned(
+            child: child,
+            bottom: padding.top > 0 ? null : 0,
+            top: padding.bottom > 0 ? null : 0,
+          )
+        ]);
+      });
+
+  static PlaceholderAlignment _tryParse(String value) {
+    switch (value) {
+      case _kCssVerticalAlignTop:
+      case _kCssVerticalAlignSub:
+        return PlaceholderAlignment.top;
+      case _kCssVerticalAlignSuper:
+      case _kCssVerticalAlignBottom:
+        return PlaceholderAlignment.bottom;
+      case _kCssVerticalAlignMiddle:
+        return PlaceholderAlignment.middle;
+    }
+
+    return null;
   }
-}
-
-_VerticalAlignInput _buildInput(String verticalAlign, TextStyleBuilders tsb) {
-  switch (verticalAlign) {
-    case _kCssVerticalAlignTop:
-      return _VerticalAlignInput(PlaceholderAlignment.top);
-    case _kCssVerticalAlignBottom:
-      return _VerticalAlignInput(PlaceholderAlignment.bottom);
-    case _kCssVerticalAlignMiddle:
-      return _VerticalAlignInput(PlaceholderAlignment.middle);
-    case _kCssVerticalAlignSub:
-      return _VerticalAlignInput(
-        PlaceholderAlignment.top,
-        padding: EdgeInsets.only(bottom: .4),
-        tsb: tsb,
-      );
-    case _kCssVerticalAlignSuper:
-      return _VerticalAlignInput(
-        PlaceholderAlignment.bottom,
-        padding: EdgeInsets.only(top: .4),
-        tsb: tsb,
-      );
-  }
-
-  return null;
-}
-
-@immutable
-class _VerticalAlignInput {
-  final PlaceholderAlignment alignment;
-  final EdgeInsets padding;
-  final TextStyleBuilders tsb;
-
-  _VerticalAlignInput(this.alignment, {this.padding, this.tsb})
-      : assert((padding == null) == (tsb == null));
 }
