@@ -21,12 +21,12 @@ class _TagA {
           for (final piece in pieces) {
             if (piece.hasWidgets) {
               for (final placeholder in piece.widgets) {
-                placeholder.wrapWith(wf.buildGestureDetectors, onTap);
+                placeholder.wrapWith((w) => wf.buildGestureDetector(w, onTap));
               }
             } else {
               for (final bit in piece.text.bits.toList(growable: false)) {
                 if (bit is TextWidget) {
-                  bit.widget.wrapWith(wf.buildGestureDetectors, onTap);
+                  bit.widget.wrapWith((w) => wf.buildGestureDetector(w, onTap));
                 } else if (bit is TextData) {
                   bit.replaceWith(_TagATextData(bit, onTap, wf));
                 }
@@ -41,8 +41,7 @@ class _TagA {
   GestureTapCallback _buildGestureTapCallback(NodeMetadata meta) {
     final attrs = meta.domElement.attributes;
     final href = attrs.containsKey('href') ? attrs['href'] : null;
-    final url = wf.constructFullUrl(href);
-    return wf.buildGestureTapCallbackForUrl(url);
+    return wf.buildGestureTapCallbackForUrl(wf.constructFullUrl(href) ?? href);
   }
 }
 
@@ -55,9 +54,10 @@ class _TagATextData extends TextData {
       : super(bit.parent, bit.data, bit.tsb);
 
   @override
-  bool get canCompile => true;
+  bool get hasBuilder => true;
 
   @override
-  InlineSpan compile(TextStyle style) =>
-      wf.buildGestureTapCallbackSpan(bit.data, onTap, style);
+  TextSpanBuilder prepareBuilder(TextStyleBuilder tsb) =>
+      TextSpanBuilder((context) => wf.buildGestureTapCallbackSpan(
+          bit.data, onTap, tsb.build(context).styleWithHeight));
 }
