@@ -1,25 +1,74 @@
+import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_widget_from_html_core/flutter_widget_from_html_core.dart';
 
 import '_.dart';
 
 void main() {
-  group('getInlineStyles', () {
+  group('defaultStyles', () {
     testWidgets('renders inline style normally', (tester) async {
       final html = '<span style="color: #f00; color: #0f0;">Foo</span>';
-      final e = await explain(tester, html);
-      expect(e, equals('[RichText:(#FF00FF00:Foo)]'));
+      final explained = await explain(tester, html);
+      expect(explained, equals('[RichText:(#FF00FF00:Foo)]'));
     });
 
-    testWidgets('renders getInlineStyles in reversed', (tester) async {
+    testWidgets('renders defaultStyles in reversed', (tester) async {
       final html = '<span>Foo</span>';
-      final e = await explain(tester, null,
+      final explained = await explain(tester, null,
           hw: HtmlWidget(
             html,
-            factoryBuilder: () => _GetInlineStylesTest(),
+            factoryBuilder: () => _DefaultStylesTest(),
             key: hwKey,
           ));
-      expect(e, equals('[RichText:(#FFFF0000:Foo)]'));
+      expect(explained, equals('[RichText:(#FFFF0000:Foo)]'));
+    });
+  });
+
+  group('onPieces', () {
+    testWidgets('renders additional text', (tester) async {
+      final html = '<span>Foo</span>';
+      final explained = await explain(tester, null,
+          hw: HtmlWidget(
+            html,
+            factoryBuilder: () => _OnPiecesTestText(),
+            key: hwKey,
+          ));
+      expect(explained, equals('[RichText:(:Foo bar)]'));
+    });
+
+    testWidgets('renders widget', (tester) async {
+      final html = '<span>Foo</span>';
+      final explained = await explain(tester, null,
+          hw: HtmlWidget(
+            html,
+            factoryBuilder: () => _OnPiecesTestWidget(),
+            key: hwKey,
+          ),
+          useExplainer: false);
+      expect(
+          explained,
+          equals('WidgetPlaceholder<Widget>\n'
+              '└Text("Hi")\n'
+              ' └RichText(text: "Hi")\n\n'));
+    });
+  });
+
+  group('onWidgets', () {
+    testWidgets('renders widget', (tester) async {
+      final html = '<span>Foo</span>';
+      final explained = await explain(tester, null,
+          hw: HtmlWidget(
+            html,
+            factoryBuilder: () => _OnWidgetsTest(),
+            key: hwKey,
+          ),
+          useExplainer: false);
+      expect(
+          explained,
+          equals('WidgetPlaceholder<BuildOp>\n'
+              '└CssBlock()\n'
+              ' └Text("Hi")\n'
+              '  └RichText(text: "Hi")\n\n'));
     });
   });
 
@@ -27,33 +76,75 @@ void main() {
     final html = '<span>Foo</span>';
 
     testWidgets('renders A first', (tester) async {
-      final e = await explain(tester, null,
+      final explained = await explain(tester, null,
           hw: HtmlWidget(
             html,
             factoryBuilder: () => _PriorityTest(a: 1, b: 2),
             key: hwKey,
           ));
-      expect(e, equals('[RichText:(:Foo A B)]'));
+      expect(explained, equals('[RichText:(:Foo A B)]'));
     });
 
     testWidgets('renders B first', (tester) async {
-      final e = await explain(tester, null,
+      final explained = await explain(tester, null,
           hw: HtmlWidget(
             html,
             factoryBuilder: () => _PriorityTest(a: 2, b: 1),
             key: hwKey,
           ));
-      expect(e, equals('[RichText:(:Foo B A)]'));
+      expect(explained, equals('[RichText:(:Foo B A)]'));
     });
   });
 }
 
-class _GetInlineStylesTest extends WidgetFactory {
+class _DefaultStylesTest extends WidgetFactory {
   @override
   void parseTag(NodeMetadata meta, String tag, Map<dynamic, String> attrs) {
     meta.op = BuildOp(defaultStyles: (_, __) => {'color': '#f00'});
     meta.op = BuildOp(defaultStyles: (_, __) => {'color': '#0f0'});
 
+    return super.parseTag(meta, tag, attrs);
+  }
+}
+
+class _OnPiecesTestText extends WidgetFactory {
+  @override
+  void parseTag(NodeMetadata meta, String tag, Map<dynamic, String> attrs) {
+    meta.op = BuildOp(onPieces: (_, pieces) {
+      for (final piece in pieces) {
+        if (piece.hasWidgets) continue;
+        piece.text.addText(' bar');
+      }
+
+      return pieces;
+    });
+    return super.parseTag(meta, tag, attrs);
+  }
+}
+
+class _OnPiecesTestWidget extends WidgetFactory {
+  @override
+  void parseTag(NodeMetadata meta, String tag, Map<dynamic, String> attrs) {
+    meta.op = BuildOp(onPieces: (_, pieces) {
+      for (final piece in pieces) {
+        if (piece.hasWidgets) continue;
+        for (final bit in List<TextBit>.unmodifiable(piece.text.bits)) {
+          bit.detach();
+        }
+      }
+
+      return [
+        BuiltPiece.widgets([Text('Hi')])
+      ];
+    });
+    return super.parseTag(meta, tag, attrs);
+  }
+}
+
+class _OnWidgetsTest extends WidgetFactory {
+  @override
+  void parseTag(NodeMetadata meta, String tag, Map<dynamic, String> attrs) {
+    meta.op = BuildOp(onWidgets: (_, __) => [Text('Hi')]);
     return super.parseTag(meta, tag, attrs);
   }
 }
