@@ -40,6 +40,7 @@ Future<String> explain(
   void Function(BuildContext) preTest,
   bool rtl = false,
   TextStyle textStyle,
+  bool useExplainer = true,
 }) async {
   assert((html == null) != (hw == null));
   hw ??= HtmlWidget(
@@ -81,6 +82,21 @@ Future<String> explain(
       },
     ),
   );
+
+  if (!useExplainer) {
+    final sb = StringBuffer();
+    hwKey.currentContext.visitChildElements(
+        (e) => sb.writeln(e.toDiagnosticsNode().toStringDeep()));
+    var str = sb.toString();
+    str = str.replaceAll(RegExp(r': [A-Z][A-Za-z]+\.'), ': '); // enums
+    str = str.replaceAll(RegExp(r'\[GlobalKey#[0-9a-f]+\]'), '');
+    str = str.replaceAll(RegExp(r'(, )?dependencies: \[[^\]]+\]'), '');
+    str = str.replaceAll(
+        RegExp(r'(, )?renderObject: .+ relayoutBoundary=\w+'), '');
+    str = str.replaceAll(RegExp(r'softWrap: [a-z\s]+, '), '');
+    str = str.replaceAll('maxLines: unlimited, ', '');
+    return str;
+  }
 
   final hws = hwKey.currentState;
   expect(hws, isNotNull);
@@ -176,12 +192,18 @@ class Explainer {
 
     final c = w.constraints;
     final s = w.size;
-    if (s.height.isFinite) attr.add('height=${s.height}');
-    if (c.maxHeight.isFinite) attr.add('maxHeight=${c.maxHeight}');
-    if (c.maxWidth.isFinite) attr.add('maxWidth=${c.maxWidth}');
-    if (c.minHeight > 0) attr.add('minHeight=${c.minHeight}');
-    if (c.minWidth > 0) attr.add('minWidth=${c.minWidth}');
-    if (s.width.isFinite) attr.add('width=${s.width}');
+    if (s.height.isFinite) attr.add('height=${s.height.toStringAsFixed(1)}');
+    if (c.maxHeight.isFinite) {
+      attr.add('maxHeight=${c.maxHeight.toStringAsFixed(1)}');
+    }
+    if (c.maxWidth.isFinite) {
+      attr.add('maxWidth=${c.maxWidth.toStringAsFixed(1)}');
+    }
+    if (c.minHeight > 0) {
+      attr.add('minHeight=${c.minHeight.toStringAsFixed(1)}');
+    }
+    if (c.minWidth > 0) attr.add('minWidth=${c.minWidth.toStringAsFixed(1)}');
+    if (s.width.isFinite) attr.add('width=${s.width.toStringAsFixed(1)}');
 
     return attr;
   }
@@ -313,7 +335,7 @@ class Explainer {
     }
 
     if (style.height != null) {
-      s += '+height=${style.height}';
+      s += '+height=${style.height.toStringAsFixed(1)}';
     }
 
     if (style.fontSize != parent.fontSize) {
@@ -377,21 +399,13 @@ class Explainer {
     // ignore: invalid_use_of_protected_member
     if (widget is WidgetPlaceholder) return _widget(widget.build(context));
 
-    if (widget is Image) return _image(widget);
+    if (widget is Builder) return _widget(widget.builder(context));
 
-    if (widget is LayoutBuilder) {
-      return _widget(widget.builder(
-        context,
-        BoxConstraints.tightFor(width: 800, height: 600),
-      ));
-    }
+    if (widget is Image) return _image(widget);
 
     if (widget is SizedBox) return _sizedBox(widget);
 
-    final type = widget.runtimeType
-        .toString()
-        .replaceAll('_MarginHorizontal', 'Padding');
-
+    final type = '${widget.runtimeType}';
     var attr = <String>[];
 
     attr.add(_textAlign(widget is RichText
